refactor: 불필요한 파일 삭제 및 패키지명 변경
