--- src/main/java/chess/dao/Role.java
@@ -1,21 +1,0 @@
-package chess.dao;
-
-public class Role {
-
-    private final String role;
-
-    public Role(String role) {
-        this.role = role;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    @Override
-    public String toString() {
-        return "Role{" +
-                "role='" + role + '\'' +
-                '}';
-    }
-}
