+++ src/main/java/chess/dto/RoomDto.java
@@ -1,4 +1,4 @@
-package chess.domain.game;
+package chess.dto;
 
 import chess.dao.Member;
